refactor(cli): minimize seller UX and use shared SDK clients
