+++ apps/cli/package.json
@@ -35,7 +35,6 @@
     "cli-table3": "^0.6.5",
     "commander": "^14.0.3",
     "dotenv": "^16.6.1",
-    "ethers": "~6.16.0",
     "open": "^11.0.0",
     "ora": "^9.3.0",
     "qrcode": "^1.5.4"
