+++ apps/cli/src/cli/commands/network/index.ts
@@ -2,7 +2,6 @@ import type { Command } from 'commander';
 import { registerNetworkBrowseCommand } from './browse.js';
 import { registerNetworkPeerCommand } from './peer.js';
 import { registerNetworkBootstrapCommand } from './bootstrap.js';
-import { registerNetworkContractsCommand } from './contracts.js';
 
 export function registerNetworkCommands(program: Command): void {
   const networkCmd = program
@@ -12,5 +11,4 @@ export function registerNetworkCommands(program: Command): void {
   registerNetworkBrowseCommand(networkCmd);
   registerNetworkPeerCommand(networkCmd);
   registerNetworkBootstrapCommand(networkCmd);
-  registerNetworkContractsCommand(networkCmd);
 }
