--- apps/cli/src/cli/commands/seller/emissions.ts
@@ -1,6 +1,0 @@
-import type { Command } from 'commander';
-import { registerEmissionsCommand } from '../emissions.js';
-
-export function registerSellerEmissionsCommand(sellerCmd: Command): void {
-  registerEmissionsCommand(sellerCmd, 'seller');
-}
