+++ apps/cli/src/cli/commands/seller/index.ts
@@ -4,7 +4,6 @@ import { registerSellerSetupCommand } from './setup.js';
 import { registerSellerStatusCommand } from './status.js';
 import { registerSellerRegisterCommand } from './register.js';
 import { registerSellerStakeCommand } from './stake.js';
-import { registerSellerEmissionsCommand } from './emissions.js';
 import { registerSellerDoctorCommand } from './doctor.js';
 import { registerSellerPoolCommand } from './pool.js';
 import { registerSellerRewardsCommand } from './rewards.js';
@@ -19,7 +18,6 @@ export function registerSellerCommands(program: Command): void {
   registerSellerStatusCommand(sellerCmd);
   registerSellerRegisterCommand(sellerCmd);
   registerSellerStakeCommand(sellerCmd);
-  registerSellerEmissionsCommand(sellerCmd);
   registerSellerPoolCommand(sellerCmd);
   registerSellerRewardsCommand(sellerCmd);
   registerSellerDoctorCommand(sellerCmd);
