+++ apps/website/docs/guides/become-a-provider.md
@@ -14,7 +14,7 @@ AntSeed is designed for providers who build differentiated services — such as 
 :::
 
 :::info Seller ANTS emissions
-The CLI verifies the active contract stack through `AntseedRegistry`. On the legacy stack, seller rewards follow the existing emissions path. After the recognized-usage cutover, finalized legacy rewards remain claimable and new seller/operator plus pool-staker rewards use the recognized-usage contracts.
+The CLI checks `AntseedRegistry` to determine whether the recognized-usage upgrade is active. Before the upgrade, seller rewards follow the existing emissions path. After it, finalized earlier rewards remain claimable and new seller/operator plus pool-staker rewards use the recognized-usage contracts.
 :::
 
 ## Prerequisites
@@ -213,7 +213,7 @@ antseed seller status
 # Register your identity on-chain (ERC-8004)
 antseed seller register
 
-# Stake USDC (minimum $10, legacy stack)
+# Stake USDC before the recognized-usage upgrade (minimum $10)
 antseed seller legacy stake 10
 
 # Verify everything is ready
@@ -224,12 +224,14 @@ On networks that have completed the recognized-usage cutover, use an ANTS seller
 
 ```bash
 antseed seller register
-antseed seller pool bootstrap
+antseed seller legacy claim-starter
 antseed seller stake 100 --epochs 4
 antseed seller pool positions
 ```
 
-Pool stake activates after the contract's activation delay. The CLI reports positions as pending, active, matured, closed, or withdrawn and refuses an early withdrawal unless `--force` is supplied to acknowledge slashing.
+`seller stake` always means ANTS; use `seller legacy stake` for USDC on networks that have not upgraded. The starter claim is only for eligible legacy sellers; new sellers register and stake ANTS without that step.
+
+Pool stake activates after the contract's activation delay. The CLI reports positions as pending, active, matured, closed, or withdrawn. For an early withdrawal, `--accept-slashing` first prints the estimated principal loss and then requires confirmation; add `--yes` only for non-interactive automation after reviewing that estimate. The rate can change before the transaction executes, so the estimate is not a guaranteed maximum loss.
 
 ## 7. Add Your Services
