+++ packages/node/src/index.ts
@@ -158,9 +158,10 @@ export { EmissionsClient, type EmissionsClientConfig, type EmissionsEpochParams 
 export { RegistryClient, type RegistryClientConfig } from './payments/evm/registry-client.js';
 export { UsageAccountingClient, type UsageAccountingClientConfig } from './payments/evm/usage-accounting-client.js';
 export { UsageRewardsClient, type UsageRewardsClientConfig } from './payments/evm/usage-rewards-client.js';
-export { SellerPoolsClient, type SellerPoolsClientConfig, type SellerPoolPosition } from './payments/evm/seller-pools-client.js';
+export { SellerPoolsClient, estimateEarlyExit, type SellerPoolsClientConfig, type SellerPoolPosition, type EarlyExitEstimate } from './payments/evm/seller-pools-client.js';
 export { SellerPoolsRewardsClient, type SellerPoolsRewardsClientConfig } from './payments/evm/seller-pools-rewards-client.js';
-export { SellerRegistryClient, type SellerRegistryClientConfig } from './payments/evm/seller-registry-client.js';
+export { SellerRegistryClient, SellerRegistrationVerificationError, type SellerRegistryClientConfig } from './payments/evm/seller-registry-client.js';
+export { pendingEpochRewards, claimEpochRewards, claimBuyerEpochRewards, previewPoolRewards, claimPoolRewards, type RewardTransactionRecorder } from './payments/reward-claims.js';
 export { PositionInitClient, type PositionInitClientConfig } from './payments/evm/position-init-client.js';
 export { EmissionsGateClient, type EmissionsGateClientConfig } from './payments/evm/emissions-gate-client.js';
 export {
