+++ packages/node/src/payments/index.ts
@@ -90,12 +90,14 @@ export { UsageAccountingClient } from './evm/usage-accounting-client.js';
 export type { UsageAccountingClientConfig } from './evm/usage-accounting-client.js';
 export { UsageRewardsClient } from './evm/usage-rewards-client.js';
 export type { UsageRewardsClientConfig } from './evm/usage-rewards-client.js';
-export { SellerPoolsClient } from './evm/seller-pools-client.js';
-export type { SellerPoolsClientConfig, SellerPoolPosition } from './evm/seller-pools-client.js';
+export { SellerPoolsClient, estimateEarlyExit } from './evm/seller-pools-client.js';
+export type { SellerPoolsClientConfig, SellerPoolPosition, EarlyExitEstimate } from './evm/seller-pools-client.js';
 export { SellerPoolsRewardsClient } from './evm/seller-pools-rewards-client.js';
 export type { SellerPoolsRewardsClientConfig } from './evm/seller-pools-rewards-client.js';
-export { SellerRegistryClient } from './evm/seller-registry-client.js';
+export { SellerRegistryClient, SellerRegistrationVerificationError } from './evm/seller-registry-client.js';
 export type { SellerRegistryClientConfig } from './evm/seller-registry-client.js';
+export { pendingEpochRewards, claimEpochRewards, claimBuyerEpochRewards, previewPoolRewards, claimPoolRewards } from './reward-claims.js';
+export type { RewardTransactionRecorder } from './reward-claims.js';
 export { PositionInitClient } from './evm/position-init-client.js';
 export type { PositionInitClientConfig } from './evm/position-init-client.js';
 export { EmissionsGateClient } from './evm/emissions-gate-client.js';
